restyle calendar page and adding bugs dark more logo
