+++ src/TeamPage.js
@@ -7,15 +7,8 @@ import eshao from "./profiles/eshao.jpg";
 import khanhng from "./profiles/khanhng.jpg";
 import slin from "./profiles/slin.jpg";
 import eriche from "./profiles/eriche.jpg";
+import Eboard from './Eboard';
 
-const fadein = keyframes`
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-`;
 
 const scalein = keyframes`
   from {
@@ -49,80 +42,6 @@ const MainDescriptionContainer = styled.section`
   margin-bottom: 5%;
 `;
 
-const DescriptionContainer = styled.div`
-  width: 100%;
-  text-align: center;
-  white-space: pre-wrap;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  padding: 5px 20px 20px 20px;
-`;
-
-const Description = styled.p`
-  text-align: left;
-  font-weight: 500;
-  font-style: normal;
-  letter-spacing: 0em;
-  text-transform: none;
-  line-height: 1.2;
-  font-size: calc(0vw + 1rem);
-  color: white;
-  animation: ${scalein} 1s;
-`;
-
-const Title = styled.h2`
-  font-weight: 300;
-  font-style: normal;
-  letter-spacing: 0em;
-  text-transform: none;
-  text-align: center;
-  line-height: 1;
-  margin-bottom: 0;
-  font-size: calc(0vw + 1.5rem);
-  color: white;
-  animation: ${scalein} 0.9s;
-`;
-
-const ProjectContainer = styled.div`
-  margin: 1vw;
-  grid-column: 1
-  min-height: 40vh;
-  height: 100%;
-  background-color: none;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 5px 60px 0px 60px;
-  animation: ${fadein} 1.5s;
-  @media screen and (max-width: 768px) {
-    padding: 0px 0px 0px 0px;
-    margin: 0 0 25px 0;
-    width: 70vw;
-  }
-`;
-
-const SubTitle = styled.p`
-  font-family: var(--normal-font);
-  text-align: center;
-  font-weight: lighter;
-  color: white;
-  margin: 10px 0 0 0;
-  animation: ${scalein} 0.9s;
-  @media screen and (max-width: 768px) {
-    font-size: 1.25rem;
-  }
-`;
-
-const Image = styled.img`
-  width: 70%;
-  object-fit: contain;
-  animation: ${fadein} 1.5s;
-  @media screen and (max-width: 768px) {
-    width: 50%;
-  }
-`;
-
 const EboardTitleContainer = styled.div`
   margin-top: 15%;
   text-align: center;
@@ -143,7 +62,7 @@ const EboardTitle = styled.h1`
   color: rgba(255, 255, 255, 0.7);
   animation: ${scalein} 0.9s;
   @media screen and (max-width: 768px) {
-    font-size: calc(0vw + 1.5rem);
+    font-size: calc(0.5vw + 1.5rem);
   }
 `;
 
@@ -201,23 +120,6 @@ const eboards = {
   },
 };
 
-const Member = (props) => {
-  const { name, role, image, description } = props;
-  return (
-    <ProjectContainer>
-      <Image src={image} />
-      <Title>
-        {" "}
-        {name}
-        <SubTitle> {role} </SubTitle>
-      </Title>
-      <DescriptionContainer>
-        <Description>{description}</Description>
-      </DescriptionContainer>
-    </ProjectContainer>
-  );
-};
-
 const TeamPage = () => {
   return (
     <MainContainer>
@@ -230,7 +132,7 @@ const TeamPage = () => {
         <MainDescriptionContainer>
           {Object.entries(eboards).map(([name, values]) => {
             return (
-              <Member
+              <Eboard
                 name={name}
                 role={values.role}
                 description={values.description}
